começo api usuario/paciente

+++ posto-saude-api/src/main/java/io/github/worthdavi/postosaude/repository/AgendaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import io.github.worthdavi.postosaude.model.Agenda;
 
 
-public interface AgendaRepository extends JpaRepository<Agenda, Long> {
+public interface AgendaRepository extends JpaRepository<Agenda, Integer> {
 	
 }
