+++ posto-saude-api/src/main/java/io/github/worthdavi/postosaude/repository/ConsultaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import io.github.worthdavi.postosaude.model.Consulta;
 
 
-public interface ConsultaRepository extends JpaRepository<Consulta, Long> {
+public interface ConsultaRepository extends JpaRepository<Consulta, Integer> {
 	
 }
